@@ -1,3 +1,3 @@
 module github.com/TheAlgorithms/Go
 
-go 1.18
+go 1.19
