@@ -9,12 +9,12 @@ func Encrypt(input string, key int) string {
 	key8 := byte(key%26+26) % 26
 
 	var outputBuffer []byte
-	// r is a rune, which is the equivalent of uint32.
-	for _, r := range input {
-		newByte := byte(r)
-		if 'A' <= r && r <= 'Z' {
+	// b is a byte, which is the equivalent of uint8.
+	for _, b := range []byte(input) {
+		newByte := b
+		if 'A' <= b && b <= 'Z' {
 			outputBuffer = append(outputBuffer, 'A'+(newByte-'A'+key8)%26)
-		} else if 'a' <= r && r <= 'z' {
+		} else if 'a' <= b && b <= 'z' {
 			outputBuffer = append(outputBuffer, 'a'+(newByte-'a'+key8)%26)
 		} else {
 			outputBuffer = append(outputBuffer, newByte)
