@@ -8,9 +8,9 @@ package nested
 //
 // A sequence of brackets `s` is considered properly nested
 // if any of the following conditions are true:
-// 	- `s` is empty;
-// 	- `s` has the form (U) or [U] or {U} where U is a properly nested string;
-// 	- `s` has the form VW where V and W are properly nested strings.
+//   - `s` is empty;
+//   - `s` has the form (U) or [U] or {U} where U is a properly nested string;
+//   - `s` has the form VW where V and W are properly nested strings.
 //
 // For example, the string "()()[()]" is properly nested but "[(()]" is not.
 //
