@@ -37,8 +37,8 @@ func BinaryIterative(array []int, target int) (int, error) {
 	return -1, ErrNotFound
 }
 
-//Returns index to the first element in the range [0, len(array)-1] that is not less than (i.e. greater or equal to) target.
-//return -1 and ErrNotFound if no such element is found.
+// Returns index to the first element in the range [0, len(array)-1] that is not less than (i.e. greater or equal to) target.
+// return -1 and ErrNotFound if no such element is found.
 func LowerBound(array []int, target int) (int, error) {
 	startIndex := 0
 	endIndex := len(array) - 1
@@ -59,8 +59,8 @@ func LowerBound(array []int, target int) (int, error) {
 	return startIndex, nil
 }
 
-//Returns index to the first element in the range [lowIndex, len(array)-1] that is greater than target.
-//return -1 and ErrNotFound if no such element is found.
+// Returns index to the first element in the range [lowIndex, len(array)-1] that is greater than target.
+// return -1 and ErrNotFound if no such element is found.
 func UpperBound(array []int, target int) (int, error) {
 	startIndex := 0
 	endIndex := len(array) - 1
