@@ -6,7 +6,7 @@ func Insertion[T constraints.Ordered](arr []T) []T {
 	for currentIndex := 1; currentIndex < len(arr); currentIndex++ {
 		temporary := arr[currentIndex]
 		iterator := currentIndex
-		for ; iterator > 0 && arr[iterator-1] >= temporary; iterator-- {
+		for ; iterator > 0 && arr[iterator-1] > temporary; iterator-- {
 			arr[iterator] = arr[iterator-1]
 		}
 		arr[iterator] = temporary
