@@ -1,2 +1,2 @@
-// Package linkedlist demonstates different implementations on linkedlists.
+// Package linkedlist demonstrates different implementations on linkedlists.
 package linkedlist
