@@ -4,7 +4,7 @@ import (
 	"testing"
 )
 
-//Query interval
+// Query interval
 type query struct {
 	firstIndex int
 	lastIndex  int
@@ -59,6 +59,14 @@ func TestSegmentTree(t *testing.T) {
 				{firstIndex: 2, lastIndex: 4, value: 2}},
 			queries:  []query{{0, 5}, {0, 2}, {2, 4}},
 			expected: []int{31, 14, 24},
+		},
+		{
+			description: "test array with size 11 and range updates",
+			array:       []int{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
+			updates: []update{{firstIndex: 2, lastIndex: 8, value: 2},
+				{firstIndex: 2, lastIndex: 8, value: 2}},
+			queries:  []query{{3, 5}, {7, 8}, {4, 5}, {8, 8}},
+			expected: []int{15, 10, 10, 5},
 		},
 	}
 	for _, test := range segmentTreeTestData {
